+++ src/NetCore/Web/Api/wwwroot/scripts/api/ui/webgis.ui.plugins.general.js
@@ -72,4 +72,4 @@ webgis.ui.showLayerNotVisibleNotification = function (service, query, $target) {
                 });
         }
     }
-}
+};