+++ src/NetCore/Web/Api/wwwroot/scripts/api/webgis.options.js
@@ -93,7 +93,7 @@ webgis.usability = {
 
     useAdvancedKeyShortcutHandling: false,
 
-    showQueryLayerNotVisbleNotification: true,
+    showQueryLayerNotVisbleNotification: false,
 
     tsp: {
         allowOrderFeatures: false,